A first version of opening index files with proper configuration (#301)
