+++ git-repository/src/repository/mod.rs
@@ -39,48 +39,7 @@ impl crate::Repository {
     }
 }
 
-mod worktree {
-    use crate::{worktree, Worktree};
-
-    impl crate::Repository {
-        /// Return a platform for interacting with worktrees
-        pub fn worktree(&self) -> worktree::Platform<'_> {
-            worktree::Platform { parent: self }
-        }
-    }
-
-    impl<'repo> worktree::Platform<'repo> {
-        /// Return the currently set worktree if there is one.
-        ///
-        /// Note that there would be `None` if this repository is `bare` and the parent [`Repository`][crate::Repository] was instantiated without
-        /// registered worktree in the current working dir.
-        pub fn current(&self) -> Option<Worktree<'repo>> {
-            self.parent.work_dir().map(|path| Worktree {
-                parent: self.parent,
-                path,
-            })
-        }
-    }
-
-    impl<'repo> Worktree<'repo> {
-        /// Open a new copy of the index file and decode it entirely.
-        ///
-        /// It will use the `index.threads` configuration key to learn how many threads to use.
-        #[cfg(feature = "git-index")]
-        pub fn open_index(&self) -> Result<git_index::File, git_index::file::init::Error> {
-            let repo = self.parent;
-            // repo.config.resolved.value::<git_config::values::Boolean>("index", None, "threads")
-            git_index::File::at(
-                repo.git_dir().join("index"),
-                git_index::decode::Options {
-                    object_hash: repo.object_hash(),
-                    thread_limit: None, // TODO: read config
-                    min_extension_block_in_bytes_for_threading: 0,
-                },
-            )
-        }
-    }
-}
+mod worktree;
 
 /// Various permissions for parts of git repositories.
 pub mod permissions;
